Rely on hardcoded defaults in engine.rb instead of metadata.yml
The documentation fields are already correct, so the only change is that
the version is now loaded from the engine file. It is changed to have a
-development suffix. The file is still loaded in non-production setups
so CI or developers can easily override this. Because of that, it's not
ignored by git.

The primary motivation for this is avoiding to ship it in packaging.