Fix french translation to state Proxy HTTP
By default it is translated and branded to `Capsule HTTP` which
is not correct.
This stems from the fact that the theme assumes branded words
always appear in the same order. In this specific case, the
assumption is broken.
This is a workaround until we find a better way to deal with
translations and branding